# alignment_openapi.yaml
openapi: 3.0.3
info:
  title: ATOM AI Backend API
  version: 1.0.0
  description: API contract for ATOM AI Workspace
paths:
...
  /api/buildings:
    get:
      operationId: buildingsList
      tags:
      - buildings
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuildingListItem'
          description: ''
  /api/buildings/{building_id}:
    get:
      operationId: buildingDetail
      parameters:
      - in: path
        name: building_id
        schema:
          type: string
        required: true
      tags:
      - buildings
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingDetail'
          description: ''
  /api/buildings/{building_id}/departments:
    get:
      operationId: buildingDepartments
      parameters:
      - in: path
        name: building_id
        schema:
          type: string
        required: true
      tags:
      - buildings
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
          description: ''
  /api/buildings/{building_id}/floors/{floor_id}/employees/{employee_id}/profile:
    get:
      operationId: employeeProfileByFloor
      parameters:
      - in: path
        name: building_id
        schema:
          type: string
        required: true
      - in: path
        name: employee_id
        schema:
          type: string
        required: true
      - in: path
        name: floor_id
        schema:
          type: string
        required: true
      tags:
      - buildings
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeProfile'
          description: ''
  /api/buildings/{building_id}/floors/{floor_id}/workspace:
    get:
      operationId: floorWorkspace
      parameters:
      - in: path
        name: building_id
        schema:
          type: string
        required: true
      - in: path
        name: floor_id
        schema:
          type: string
        required: true
      tags:
      - buildings
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: ''
  /api/buildings/{building_id}/floors/{floor_id}/workspace-context:
    get:
      operationId: workspaceContextAlias
      parameters:
      - in: path
        name: building_id
        schema:
          type: string
        required: true
      - in: query
        name: employee_id
        schema:
          type: string
        required: true
      - in: path
        name: floor_id
        schema:
          type: string
        required: true
      tags:
      - buildings
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeWorkspaceContext'
          description: ''
  /api/buildings/{building_id}/floors/{floor_id}/workspace/employee/{employee_id}:
    get:
      operationId: employeeWorkspaceContext
      parameters:
      - in: path
        name: building_id
        schema:
          type: string
        required: true
      - in: path
        name: employee_id
        schema:
          type: string
        required: true
      - in: path
        name: floor_id
        schema:
          type: string
        required: true
      tags:
      - buildings
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeWorkspaceContext'
          description: ''
...
  /api/employees/{employee_id}/profile:
    get:
      operationId: employeeProfileGlobalAlias
      parameters:
      - in: query
        name: building_id
        schema:
          type: string
        required: true
      - in: path
        name: employee_id
        schema:
          type: string
        required: true
      - in: query
        name: floor_id
        schema:
          type: string
        required: true
      tags:
      - employees
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeProfile'
          description: ''
...
  /api/workspace/context:
    get:
      operationId: workspaceContextGlobalAlias
      parameters:
      - in: query
        name: building_id
        schema:
          type: string
        required: true
      - in: query
        name: employee_id
        schema:
          type: string
        required: true
      - in: query
        name: floor_id
        schema:
          type: string
        required: true
      tags:
      - workspace
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeWorkspaceContext'
          description: ''
components:
  schemas:
    Activity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        timestamp:
          type: string
      required:
      - id
      - timestamp
      - title
      - type
...
    BuildingDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        departments:
          type: array
          items:
            $ref: '#/components/schemas/Department'
      required:
      - departments
      - id
      - name
    BuildingListItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        floors:
          type: integer
        employees:
          type: integer
        online_now:
          type: integer
        risky_tasks:
          type: integer
        overdue_tasks:
          type: integer
        color:
          type: string
        height_ratio:
          type: number
          format: double
        latest_event:
          type: string
        logo:
          type: string
      required:
      - color
      - employees
      - floors
      - height_ratio
      - id
      - latest_event
      - logo
      - name
      - online_now
      - overdue_tasks
      - risky_tasks
      - status
...
    Comment:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        text:
          type: string
        created_at:
          type: string
      required:
      - author
      - created_at
      - id
      - text
...
    Department:
      type: object
      properties:
        id:
          type: string
        floor:
          type: integer
        name:
          type: string
        status:
          type: string
        employee_count:
          type: integer
        online_count:
          type: integer
        risky_tasks:
          type: integer
        overdue_tasks:
          type: integer
        lead:
          type: string
      required:
      - employee_count
      - floor
      - id
      - lead
      - name
      - online_count
      - overdue_tasks
      - risky_tasks
      - status
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        updated_at:
          type: string
        owner:
          type: string
      required:
      - id
      - owner
      - title
      - type
      - updated_at
    Employee:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        role:
          type: string
        avatar:
          type: string
        status:
          type: string
        email:
          type: string
        telegram:
          type: string
        timezone:
          type: string
        hours:
          type: string
        x:
          type: integer
        y:
          type: integer
        points:
          type: integer
        next_goal:
          type: integer
        next_goal_label:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceTask'
      required:
      - avatar
      - email
      - full_name
      - hours
      - id
      - next_goal
      - next_goal_label
      - points
      - role
      - status
      - tasks
      - telegram
      - timezone
      - x
      - y
    EmployeeProfile:
      type: object
      properties:
        employee:
          $ref: '#/components/schemas/Employee'
        building_name:
          type: string
        department:
          type: string
        projects:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceProject'
        activity_feed:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        comments_history:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        performance:
          $ref: '#/components/schemas/Performance'
      required:
      - activity_feed
      - building_name
      - comments_history
      - department
      - employee
      - performance
      - projects
    EmployeeWorkspaceContext:
      type: object
      properties:
        employee:
          $ref: '#/components/schemas/Employee'
        my_tasks:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceTask'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        activity_feed:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        project_context:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceProject'
      required:
      - activity_feed
      - documents
      - employee
      - my_tasks
      - project_context
...
    Performance:
      type: object
      properties:
        completed_tasks:
          type: integer
        total_tasks:
          type: integer
        on_time_rate:
          type: integer
        response_rate:
          type: integer
      required:
      - completed_tasks
      - on_time_rate
      - response_rate
      - total_tasks
...
    Workspace:
      type: object
      properties:
        building_name:
          type: string
        department:
          type: string
        zones:
          type: array
          items:
            $ref: '#/components/schemas/Zone'
        employees:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
      required:
      - building_name
      - department
      - employees
      - zones
    WorkspaceProject:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        summary:
          type: string
      required:
      - id
      - name
      - status
      - summary
    WorkspaceTask:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        column:
          type: string
        priority:
          type: string
        due:
          type: string
      required:
      - column
      - due
      - id
      - priority
      - title
    Zone:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
        w:
          type: integer
        h:
          type: integer
        label:
          type: string
        color:
          type: string
        stroke_color:
          type: string
      required:
      - color
      - h
      - label
      - stroke_color
      - w
      - x
      - y
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid